refactor(shell): split shell_cmds.c into category modules using registry table pattern
Replace the monolithic shell_cmds.c + switch/case dispatch with a
command registry approach: each category file exports a NULL-terminated
shell_cmd_entry_t[] table; shell.c iterates all modules for dispatch.

New files:
  shell_cmd_display.c  – clear, fgcol, bgcol, setmode (+ colour palette)
  shell_cmd_system.c   – echo, meminfo, uptime, tasks, shutdown, reboot, panic, ktest
  shell_cmd_disk.c     – lsdisks, lspart, mkpart, readsector, chainload
  shell_cmd_fs.c       – mount, umount, ls, cat, cd, mkdir, mkfs, isols, write, touch, cp
  shell_cmd_apps.c     – vics, install, exec, eject, ring3test

Adding a new command now only requires editing one category file —
no enum entry, no cmd_table row, no switch case, no shell_priv.h
declaration needed.

Also includes: white-on-blue default colours (VGA + VESA), fgcol/bgcol
commands, vesa_colour ktest suite, 5-second splash delay, kprint_ok blue
badge, SYS_DEBUG serial fix, docker-ktest-gui.sh macOS timeout fix.
