+++ src/kernel/arch/i386/make.config
@@ -39,5 +39,9 @@ $(ARCHDIR)/proc/installer.o \
 $(ARCHDIR)/proc/vics.o \
 $(ARCHDIR)/shell/shell.o \
 $(ARCHDIR)/shell/shell_help.o \
-$(ARCHDIR)/shell/shell_cmds.o \
+$(ARCHDIR)/shell/shell_cmd_display.o \
+$(ARCHDIR)/shell/shell_cmd_system.o \
+$(ARCHDIR)/shell/shell_cmd_disk.o \
+$(ARCHDIR)/shell/shell_cmd_fs.o \
+$(ARCHDIR)/shell/shell_cmd_apps.o \
 $(ARCHDIR)/debug/debug.o \
