--- src/kernel/arch/i386/shell/shell_cmds.c
@@ -1,1193 +1,0 @@
-/*
- * shell_cmds.c -- built-in shell command implementations.
- *
- * Every command handler declared in shell_priv.h (except cmd_help and
- * cmd_version, which live in shell_help.c) is implemented here.
- */
-
-#include "shell_priv.h"
-
-#include <kernel/tty.h>
-#include <kernel/timer.h>
-#include <kernel/heap.h>
-#include <kernel/vesa.h>
-#include <kernel/vesa_tty.h>
-#include <kernel/bochs_vbe.h>
-#include <kernel/ide.h>
-#include <kernel/partition.h>
-#include <kernel/fat32.h>
-#include <kernel/vfs.h>
-#include <kernel/iso9660.h>
-#include <kernel/installer.h>
-#include <kernel/task.h>
-#include <kernel/acpi.h>
-#include <kernel/vics.h>
-#include <kernel/debug.h>
-#include <kernel/chainload.h>
-#include <kernel/elf.h>
-#include <kernel/task.h>
-#include <kernel/vfs.h>
-
-#include <string.h>
-#include <stdint.h>
-
-/* ---------------------------------------------------------------------------
- * Private helpers
- * --------------------------------------------------------------------------- */
-
-/* Parse a simple decimal or 0x-prefixed hex number from a string. */
-static uint32_t parse_uint(const char *s)
-{
-    if (s[0] == '0' && (s[1] == 'x' || s[1] == 'X')) {
-        uint32_t v = 0;
-        s += 2;
-        while (*s) {
-            char c = *s++;
-            if (c >= '0' && c <= '9')      v = v * 16 + (uint32_t)(c - '0');
-            else if (c >= 'a' && c <= 'f') v = v * 16 + (uint32_t)(c - 'a' + 10);
-            else if (c >= 'A' && c <= 'F') v = v * 16 + (uint32_t)(c - 'A' + 10);
-            else break;
-        }
-        return v;
-    }
-    uint32_t v = 0;
-    while (*s >= '0' && *s <= '9')
-        v = v * 10 + (uint32_t)(*s++ - '0');
-    return v;
-}
-
-/* Simple hex dump of a 512-byte buffer: 32 rows of 16 bytes. */
-static void hexdump_sector(const uint8_t *buf)
-{
-    static const char hex[] = "0123456789ABCDEF";
-    for (int row = 0; row < 32; row++) {
-        int offset = row * 16;
-        /* Offset */
-        t_putchar(hex[(offset >> 8) & 0xF]);
-        t_putchar(hex[(offset >> 4) & 0xF]);
-        t_putchar(hex[(offset     ) & 0xF]);
-        t_putchar('0');
-        t_writestring(":  ");
-
-        for (int col = 0; col < 16; col++) {
-            uint8_t b = buf[offset + col];
-            t_putchar(hex[b >> 4]);
-            t_putchar(hex[b & 0xF]);
-            t_putchar(' ');
-        }
-        t_putchar('\n');
-    }
-}
-
-/* ---------------------------------------------------------------------------
- * Shared partition table storage for lspart / mkpart.
- * Stored as a static so it never lands on the kernel stack.
- * --------------------------------------------------------------------------- */
-static disk_parts_t s_disk_parts;
-
-/* Print one probed disk_parts_t to the terminal. */
-static void disk_parts_print(const disk_parts_t *dp)
-{
-    static const char hex[] = "0123456789ABCDEF";
-
-    if (dp->scheme == PART_SCHEME_NONE) {
-        t_writestring("No partition table found on this drive.\n");
-        return;
-    }
-
-    t_writestring("Scheme: ");
-    t_writestring(dp->scheme == PART_SCHEME_GPT ? "GPT" : "MBR");
-    t_writestring("  total sectors: ");
-    t_dec(dp->total_sectors);
-    t_writestring("  (");
-    t_dec(dp->total_sectors / 2048);
-    t_writestring(" MiB)\n");
-
-    if (dp->count == 0) {
-        t_writestring("No partitions.\n");
-        return;
-    }
-
-    for (int i = 0; i < dp->count; i++) {
-        const part_info_t *p = &dp->parts[i];
-
-        t_writestring("  [");
-        t_dec((uint32_t)(i + 1));
-        t_writestring("] ");
-
-        if (dp->scheme == PART_SCHEME_MBR) {
-            t_writestring("type=0x");
-            t_putchar(hex[p->mbr_type >> 4]);
-            t_putchar(hex[p->mbr_type & 0xF]);
-            t_writestring(" (");
-            t_writestring(part_type_name(p->mbr_type));
-            t_writestring(")");
-        } else {
-            t_writestring(part_guid_type_name(p->type_guid));
-            if (p->name[0] != '\0') {
-                t_writestring("  \"");
-                t_writestring(p->name);
-                t_putchar('"');
-            }
-        }
-
-        t_writestring("  LBA=");
-        t_dec(p->lba_start);
-        t_writestring("  sectors=");
-        t_dec(p->lba_count);
-        t_writestring("  size=");
-        t_dec(p->lba_count / 2048);
-        t_writestring(" MiB");
-
-        if (dp->scheme == PART_SCHEME_MBR && p->bootable)
-            t_writestring("  [boot]");
-
-        t_putchar('\n');
-    }
-}
-
-/* ---------------------------------------------------------------------------
- * Built-in command handlers
- * --------------------------------------------------------------------------- */
-
-void cmd_clear(void)
-{
-    /*
-     * Reset both outputs and restore the Medli-compatible colour scheme
-     * (white on blue).  terminal_set_colorscheme() fills the VGA buffer
-     * with the new background colour; vesa_tty_setcolor() + vesa_tty_clear()
-     * do the same for the VESA framebuffer.
-     */
-    terminal_set_colorscheme(SHELL_COLOR_VGA);
-    if (vesa_tty_is_ready()) {
-        vesa_tty_setcolor(SHELL_FG_RGB, SHELL_BG_RGB);
-        vesa_tty_clear();
-    }
-}
-
-void cmd_echo(int argc, char **argv)
-{
-    for (int i = 1; i < argc; i++) {
-        if (i > 1)
-            t_putchar(' ');
-        t_writestring(argv[i]);
-    }
-    t_putchar('\n');
-}
-
-void cmd_meminfo(void)
-{
-    t_writestring("heap used: ");
-    t_dec((uint32_t)heap_used());
-    t_writestring(" bytes\n");
-    t_writestring("heap free: ");
-    t_dec((uint32_t)heap_free());
-    t_writestring(" bytes\n");
-}
-
-void cmd_uptime(void)
-{
-    t_writestring("uptime: ");
-    t_dec(timer_get_ticks());
-    t_writestring(" ticks\n");
-}
-
-void cmd_tasks(void)
-{
-    static const char * const state_names[] = { "ready", "running", "dead" };
-    int n = task_count();
-
-    t_writestring("Tasks (");
-    t_dec((uint32_t)n);
-    t_writestring(" total):\n");
-
-    for (int i = 0; i < n; i++) {
-        task_t *t = task_get(i);
-        if (!t)
-            continue;
-        t_writestring("  [");
-        t_dec((uint32_t)i);
-        t_writestring("] ");
-        t_writestring(t->name ? t->name : "(unnamed)");
-        t_writestring(" - ");
-        t_writestring(state_names[t->state]);
-        t_putchar('\n');
-    }
-}
-
-void cmd_lsdisks(void)
-{
-    static const char * const type_str[] = { "none", "ATA", "ATAPI" };
-    int found = 0;
-
-    for (uint8_t i = 0; i < IDE_MAX_DRIVES; i++) {
-        const ide_drive_t *d = ide_get_drive(i);
-        if (!d || !d->present)
-            continue;
-
-        found = 1;
-        t_writestring("drive ");
-        t_dec(i);
-        t_writestring(": [");
-        t_writestring(d->channel == 0 ? "primary" : "secondary");
-        t_putchar(' ');
-        t_writestring(d->drive == 0 ? "master" : "slave");
-        t_writestring("] ");
-        t_writestring(type_str[d->type]);
-        t_writestring("  ");
-        t_dec(d->size / 2048);   /* MiB: sectors * 512 / (1024*1024) */
-        t_writestring(" MiB  \"");
-        t_writestring(d->model);
-        t_writestring("\"\n");
-    }
-
-    if (!found)
-        t_writestring("No drives detected.\n");
-}
-
-/* lspart <drive> — probe and display partition table (MBR or GPT). */
-void cmd_lspart(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: lspart <drive>\n");
-        return;
-    }
-
-    uint8_t drive = (uint8_t)parse_uint(argv[1]);
-
-    int err = part_probe(drive, &s_disk_parts);
-    if (err) {
-        t_writestring("Error reading drive ");
-        t_dec(drive);
-        t_writestring(": ");
-        t_dec((uint32_t)err);
-        t_putchar('\n');
-        return;
-    }
-
-    disk_parts_print(&s_disk_parts);
-}
-
-/* ---------------------------------------------------------------------------
- * mkpart helpers
- * --------------------------------------------------------------------------- */
-
-/* Static entries array for mkpart — avoids putting ~10 KB on the stack. */
-static part_info_t s_mkpart_entries[128];
-
-/* Map a type keyword to an MBR partition type byte.  Returns 0 on failure. */
-static uint8_t mkpart_mbr_type(const char *kw)
-{
-    if (strcmp(kw, "fat32") == 0) return PART_MBR_FAT32_LBA;
-    if (strcmp(kw, "mdfs")  == 0) return PART_MBR_MDFS;
-    if (strcmp(kw, "linux") == 0) return PART_MBR_LINUX;
-    if (strcmp(kw, "efi")   == 0) return PART_MBR_EFI;
-    return 0;
-}
-
-/* Map a type keyword to a GPT type GUID.  Returns 0 on success, -1 on error. */
-static int mkpart_gpt_type(const char *kw, uint8_t *out_guid)
-{
-    if (strcmp(kw, "fat32") == 0) { memcpy(out_guid, PART_GUID_FAT32, 16); return 0; }
-    if (strcmp(kw, "mdfs")  == 0) { memcpy(out_guid, PART_GUID_MDFS,  16); return 0; }
-    if (strcmp(kw, "linux") == 0) { memcpy(out_guid, PART_GUID_LINUX, 16); return 0; }
-    if (strcmp(kw, "efi")   == 0) { memcpy(out_guid, PART_GUID_EFI,   16); return 0; }
-    return -1;
-}
-
-/*
- * mkpart <drive> <mbr|gpt>
- *
- * Interactive partition table creator.  Asks the user for partition count,
- * type and size; then writes MBR or GPT structure to the chosen drive.
- *
- * Supported type keywords: fat32  mdfs  linux  efi
- * Sizes are entered in MiB; they are converted to 512-byte sectors (*2048).
- *
- * MBR layout: first partition starts at LBA 2048 (1 MiB aligned).
- * GPT layout: first usable LBA is 34; partitions aligned to LBA 2048.
- */
-void cmd_mkpart(int argc, char **argv)
-{
-    static char inbuf[64];
-
-    if (argc < 3) {
-        t_writestring("Usage: mkpart <drive> <mbr|gpt>\n");
-        return;
-    }
-
-    uint8_t drive = (uint8_t)parse_uint(argv[1]);
-
-    const ide_drive_t *drv = ide_get_drive(drive);
-    if (!drv || !drv->present) {
-        t_writestring("Error: drive not present.\n");
-        return;
-    }
-    if (drv->type != IDE_TYPE_ATA) {
-        t_writestring("Error: drive is not ATA.\n");
-        return;
-    }
-
-    int is_gpt = (strcmp(argv[2], "gpt") == 0);
-    int is_mbr = (strcmp(argv[2], "mbr") == 0);
-    if (!is_gpt && !is_mbr) {
-        t_writestring("Error: scheme must be 'mbr' or 'gpt'.\n");
-        return;
-    }
-
-    int max_parts = is_mbr ? 4 : 128;
-
-    /* ---------- ask partition count ---------- */
-    int num_parts = 0;
-    do {
-        t_writestring("Number of partitions (1–");
-        t_dec((uint32_t)max_parts);
-        t_writestring("): ");
-        shell_readline(inbuf, sizeof(inbuf));
-        num_parts = (int)parse_uint(inbuf);
-    } while (num_parts < 1 || num_parts > max_parts);
-
-    /*
-     * Available space for partitions.
-     * MBR: first partition at LBA 2048, last usable = disk_sectors − 1.
-     * GPT: first partition at LBA 2048 (first usable is 34, we align to 2048),
-     *      last usable LBA = disk_sectors − 34.
-     */
-    uint32_t disk_sectors = drv->size;
-    uint32_t next_lba     = 2048u;
-    uint32_t last_lba     = is_gpt ? disk_sectors - 34u : disk_sectors - 1u;
-
-    memset(s_mkpart_entries, 0, sizeof(s_mkpart_entries));
-
-    for (int i = 0; i < num_parts; i++) {
-        t_writestring("\nPartition ");
-        t_dec((uint32_t)(i + 1));
-        t_writestring(":\n");
-
-        /* ---------- type ---------- */
-        uint8_t  mbr_type = 0;
-        uint8_t  gpt_guid[16];
-        memset(gpt_guid, 0, 16);
-
-        while (1) {
-            t_writestring("  Type [fat32/mdfs/linux");
-            if (is_gpt) t_writestring("/efi");
-            t_writestring("]: ");
-            shell_readline(inbuf, sizeof(inbuf));
-
-            if (is_mbr) {
-                mbr_type = mkpart_mbr_type(inbuf);
-                if (mbr_type != 0) break;
-            } else {
-                if (mkpart_gpt_type(inbuf, gpt_guid) == 0) break;
-            }
-            t_writestring("  Unknown type — try again.\n");
-        }
-
-        /* ---------- size ---------- */
-        uint32_t avail_mib = (last_lba - next_lba + 1u) / 2048u;
-        /* Leave room for at least 1 MiB per remaining partition. */
-        if (avail_mib > (uint32_t)(num_parts - i))
-            avail_mib -= (uint32_t)(num_parts - i - 1);
-
-        uint32_t size_mib = 0;
-        while (size_mib < 1 || size_mib > avail_mib) {
-            t_writestring("  Size in MiB (1–");
-            t_dec(avail_mib);
-            t_writestring("): ");
-            shell_readline(inbuf, sizeof(inbuf));
-            size_mib = parse_uint(inbuf);
-        }
-
-        uint32_t lba_count = size_mib * 2048u;
-
-        /* ---------- optional name (GPT only) ---------- */
-        if (is_gpt) {
-            t_writestring("  Partition name (Enter for none): ");
-            shell_readline(inbuf, sizeof(inbuf));
-        }
-
-        /* ---------- fill in descriptor ---------- */
-        part_info_t *p = &s_mkpart_entries[i];
-        p->scheme    = is_gpt ? PART_SCHEME_GPT : PART_SCHEME_MBR;
-        p->mbr_type  = mbr_type;
-        p->bootable  = 0;
-        p->lba_start = next_lba;
-        p->lba_count = lba_count;
-        if (is_gpt) {
-            memcpy(p->type_guid, gpt_guid, 16);
-            /* part_guid left as zero; part_write_gpt() will generate one. */
-            /* Copy at most 36 characters of the partition name. */
-            for (int k = 0; k < 36 && inbuf[k] != '\0'; k++)
-                p->name[k] = inbuf[k];
-            p->name[36] = '\0';
-        }
-
-        next_lba += lba_count;
-        /* Align the start of the next partition to a 2048-sector boundary. */
-        if (next_lba % 2048u)
-            next_lba = (next_lba / 2048u + 1u) * 2048u;
-    }
-
-    /* ---------- write ---------- */
-    t_writestring("\nWriting ");
-    t_writestring(is_gpt ? "GPT" : "MBR");
-    t_writestring(" to drive ");
-    t_dec(drive);
-    t_writestring("...\n");
-
-    int err;
-    if (is_mbr)
-        err = part_write_mbr(drive, s_mkpart_entries, num_parts);
-    else
-        err = part_write_gpt(drive, s_mkpart_entries, num_parts);
-
-    if (err) {
-        t_writestring("Error writing partition table: ");
-        t_dec((uint32_t)err);
-        t_putchar('\n');
-        return;
-    }
-
-    t_writestring("Done.\n\n");
-
-    /* Verify by re-reading and displaying the result. */
-    err = part_probe(drive, &s_disk_parts);
-    if (err == 0)
-        disk_parts_print(&s_disk_parts);
-}
-
-static uint8_t sector_buf[512];
-
-void cmd_readsector(int argc, char **argv)
-{
-    if (argc < 3) {
-        t_writestring("Usage: readsector <drive> <lba>\n");
-        return;
-    }
-
-    uint8_t  drive = (uint8_t)parse_uint(argv[1]);
-    uint32_t lba   = parse_uint(argv[2]);
-
-    const ide_drive_t *d = ide_get_drive(drive);
-    if (!d || !d->present) {
-        t_writestring("Error: drive not present.\n");
-        return;
-    }
-    if (d->type != IDE_TYPE_ATA) {
-        t_writestring("Error: drive is not ATA (read not supported).\n");
-        return;
-    }
-
-    int err = ide_read_sectors(drive, lba, 1, sector_buf);
-    if (err) {
-        t_writestring("Read error: ");
-        t_dec((uint32_t)err);
-        t_putchar('\n');
-        return;
-    }
-
-    t_writestring("Sector ");
-    t_dec(lba);
-    t_writestring(" of drive ");
-    t_dec(drive);
-    t_writestring(":\n");
-    hexdump_sector(sector_buf);
-}
-
-/* ---------------------------------------------------------------------------
- * setmode — switch display mode on the fly.
- *
- * VGA text modes (Bochs VBE disabled, QEMU reverts to mode 3):
- *   80x25  /  text   — 80×25 VGA text (9×16 px glyphs)
- *   80x50            — 80×50 VGA text (9×8 px glyphs)
- *
- * VESA framebuffer modes (set via Bochs VBE I/O ports):
- *   320x240          — 320×240×32
- *   640x480  / 480p  — 640×480×32
- *   1280x720 / 720p  — 1280×720×32
- *   1920x1080/ 1080p — 1920×1080×32
- * --------------------------------------------------------------------------- */
-
-typedef struct { const char *name; uint32_t w; uint32_t h; } vesa_mode_t;
-
-static const vesa_mode_t vesa_modes[] = {
-    { "320x240",  320,  240 },
-    { "640x480",  640,  480 },
-    { "480p",     640,  480 },
-    { "1280x720", 1280, 720 },
-    { "720p",     1280, 720 },
-    { "1920x1080",1920, 1080 },
-    { "1080p",    1920, 1080 },
-};
-#define VESA_MODE_COUNT ((uint32_t)(sizeof(vesa_modes) / sizeof(vesa_modes[0])))
-
-void cmd_setmode(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: setmode <mode>\n");
-        t_writestring("  Text : 80x25  80x50\n");
-        t_writestring("  VESA : 320x240  640x480  480p  720p  1080p\n");
-        return;
-    }
-
-    const char *mode = argv[1];
-
-    /* --- VGA text modes -------------------------------------------------- */
-    if (strcmp(mode, "80x25") == 0 || strcmp(mode, "text") == 0 ||
-        strcmp(mode, "25")    == 0) {
-        bochs_vbe_disable();
-        vesa_disable();
-        vesa_tty_disable();
-        terminal_set_rows(25);
-        terminal_set_colorscheme(SHELL_COLOR_VGA);
-        t_writestring("Mode: VGA 80x25 text\n");
-        return;
-    }
-
-    if (strcmp(mode, "80x50") == 0 || strcmp(mode, "50") == 0) {
-        bochs_vbe_disable();
-        vesa_disable();
-        vesa_tty_disable();
-        terminal_set_rows(50);
-        terminal_set_colorscheme(SHELL_COLOR_VGA);
-        t_writestring("Mode: VGA 80x50 text\n");
-        return;
-    }
-
-    /* --- VESA framebuffer modes ------------------------------------------ */
-    if (!bochs_vbe_available()) {
-        t_writestring("Error: Bochs VBE not available on this hardware.\n");
-        return;
-    }
-
-    for (uint32_t i = 0; i < VESA_MODE_COUNT; i++) {
-        if (strcmp(mode, vesa_modes[i].name) != 0)
-            continue;
-
-        uint32_t w = vesa_modes[i].w;
-        uint32_t h = vesa_modes[i].h;
-
-        vesa_tty_set_scale(w >= 1280 ? 2 : 1); /* set scale before init uses it */
-        bochs_vbe_set_mode(w, h, 32);
-        vesa_update_geometry(w, h, 32);         /* fb struct reflects new size   */
-        vesa_tty_init();                        /* maps new fb region, then reinit */
-        vesa_tty_setcolor(SHELL_FG_RGB, SHELL_BG_RGB);
-        vesa_tty_clear();
-
-        t_writestring("Mode: ");
-        t_dec(w); t_writestring("x"); t_dec(h); t_writestring("x32\n");
-        return;
-    }
-
-    t_writestring("Error: unknown mode '");
-    t_writestring(mode);
-    t_writestring("'\nUsage: setmode <80x25|80x50|320x240|640x480|480p|720p|1080p>\n");
-}
-
-void cmd_panic(int argc, char **argv)
-{
-    const char *msg = (argc >= 2) ? argv[1] : "Shell-requested panic";
-    KPANIC(msg); /* never returns */
-}
-
-void cmd_shutdown(void)
-{
-    acpi_shutdown(); /* never returns */
-}
-
-void cmd_reboot(void)
-{
-    acpi_reboot(); /* never returns */
-}
-
-/* ---------------------------------------------------------------------------
- * chainload – hand off to a boot sector on disk
- *
- * Usage: chainload <drive> [lba]
- *   drive : ATA drive index (0–3)
- *   lba   : sector to read (default 0 = MBR / installed bootloader)
- *
- * Examples:
- *   chainload 0        – re-run whatever bootloader is in the MBR (GRUB, etc.)
- *   chainload 0 2048   – boot the VBR of the first 1 MiB-aligned partition
- * --------------------------------------------------------------------------- */
-
-/* 512-byte sector staging buffer — static so it doesn't live on the stack. */
-static uint8_t s_chainload_sector[512];
-
-void cmd_chainload(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: chainload <drive> [lba]\n");
-        t_writestring("  lba defaults to 0 (MBR / bootloader sector)\n");
-        return;
-    }
-
-    uint8_t  drive = (uint8_t)parse_uint(argv[1]);
-    uint32_t lba   = (argc >= 3) ? parse_uint(argv[2]) : 0;
-
-    const ide_drive_t *drv = ide_get_drive(drive);
-    if (!drv || !drv->present) {
-        t_writestring("chainload: drive not found\n");
-        return;
-    }
-
-    if (ide_read_sectors(drive, lba, 1, s_chainload_sector) != 0) {
-        t_writestring("chainload: read error\n");
-        return;
-    }
-
-    if (s_chainload_sector[510] != 0x55 || s_chainload_sector[511] != 0xAA) {
-        t_writestring("chainload: sector has no boot signature (0x55AA)\n");
-        return;
-    }
-
-    /* Copy boot sector to the BIOS-standard load address 0x7C00.
-     * The kernel's 256 MiB identity window covers 0x7C00, so phys == virt. */
-    uint8_t *dest = (uint8_t *)0x7C00;
-    for (int i = 0; i < 512; i++)
-        dest[i] = s_chainload_sector[i];
-
-    t_writestring("chainload: handing off to drive 0x");
-    t_hex(0x80 + drive);
-    t_writestring(" LBA ");
-    t_dec(lba);
-    t_writestring("\n");
-
-    chainload_enter((uint8_t)(0x80 + drive));   /* never returns */
-}
-
-/* ---------------------------------------------------------------------------
- * FAT32 filesystem commands
- * --------------------------------------------------------------------------- */
-
-/* Shared static partition table used by mount/mkfs (avoids stack pressure). */
-static disk_parts_t s_cmd_parts;
-
-/*
- * mount <drive> <part>
- *
- * Probe the partition table on <drive> and mount partition number <part>
- * (1-based, matching lspart output) as a FAT32 volume accessible at /hd/.
- */
-void cmd_mount(int argc, char **argv)
-{
-    if (argc < 3) {
-        t_writestring("Usage: mount <drive> <part>\n");
-        return;
-    }
-
-    uint8_t drive    = (uint8_t)parse_uint(argv[1]);
-    int     part_num = (int)parse_uint(argv[2]);  /* 1-based, matches lspart */
-
-    int err = part_probe(drive, &s_cmd_parts);
-    if (err) {
-        t_writestring("mount: drive not accessible\n");
-        return;
-    }
-
-    int part_idx = part_num - 1;   /* convert to 0-based array index */
-    if (part_idx < 0 || part_idx >= s_cmd_parts.count) {
-        t_writestring("mount: invalid partition number");
-        if (s_cmd_parts.count > 0) {
-            t_writestring(" (valid: 1–");
-            t_dec((uint32_t)s_cmd_parts.count);
-            t_putchar(')');
-        }
-        t_writestring("\n       (use lspart ");
-        t_dec(drive);
-        t_writestring(" to list partitions)\n");
-        return;
-    }
-
-    uint32_t lba = s_cmd_parts.parts[part_idx].lba_start;
-
-    err = fat32_mount(drive, lba);
-    if (err) {
-        t_writestring("mount: not a valid FAT32 volume (error ");
-        t_dec((uint32_t)(-err));
-        t_writestring(")\n");
-        return;
-    }
-
-    vfs_notify_hd_mounted();
-
-    t_writestring("Mounted FAT32  drive ");
-    t_dec(drive);
-    t_writestring("  partition ");
-    t_dec((uint32_t)part_num);
-    t_writestring("  LBA ");
-    t_dec(lba);
-    t_writestring("\ncwd: ");
-    t_writestring(vfs_getcwd());
-    t_putchar('\n');
-}
-
-/* umount — unmount the current FAT32 volume. */
-void cmd_umount(void)
-{
-    if (!fat32_mounted()) {
-        t_writestring("umount: no volume mounted\n");
-        return;
-    }
-    fat32_unmount();
-    vfs_notify_hd_unmounted();
-    t_writestring("Volume unmounted.\n");
-}
-
-/* ls [path] — list directory via the VFS (supports /hd/… and /cdrom/…). */
-void cmd_ls(int argc, char **argv)
-{
-    const char *path = (argc >= 2) ? argv[1] : NULL;
-
-    t_writestring(path ? path : vfs_getcwd());
-    t_writestring(":\n");
-
-    vfs_ls(path);
-}
-
-/* cat <file> — read and display a file via the VFS (supports /hd/… and /cdrom/…). */
-void cmd_cat(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: cat <file>\n");
-        return;
-    }
-    vfs_cat(argv[1]);
-}
-
-/* cd <path> — change working directory via the VFS. */
-void cmd_cd(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring(vfs_getcwd());
-        t_putchar('\n');
-        return;
-    }
-    if (vfs_cd(argv[1]) == 0) {
-        t_writestring(vfs_getcwd());
-        t_putchar('\n');
-    }
-}
-
-/* mkdir <path> — create a directory via the VFS (FAT32 /hd/ only). */
-void cmd_mkdir(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: mkdir <path>\n");
-        return;
-    }
-
-    int err = vfs_mkdir(argv[1]);
-    switch (err) {
-    case  0: t_writestring("Directory created.\n");          break;
-    case -1: t_writestring("mkdir: path error\n");           break;
-    case -2: t_writestring("mkdir: I/O error\n");            break;
-    case -4: t_writestring("mkdir: disk full\n");            break;
-    case -6: t_writestring("mkdir: already exists\n");       break;
-    default:
-        if (err < 0) {
-            t_writestring("mkdir: error ");
-            t_dec((uint32_t)(-err));
-            t_putchar('\n');
-        }
-        break;
-    }
-}
-
-/*
- * mkfs <drive> <part>
- *
- * Format partition <part> on <drive> as FAT32.
- * <part> is 1-based to match lspart output (use mkpart first).
- */
-void cmd_mkfs(int argc, char **argv)
-{
-    if (argc < 3) {
-        t_writestring("Usage: mkfs <drive> <part>\n");
-        return;
-    }
-
-    uint8_t drive    = (uint8_t)parse_uint(argv[1]);
-    int     part_num = (int)parse_uint(argv[2]);  /* 1-based, matches lspart */
-
-    int err = part_probe(drive, &s_cmd_parts);
-    if (err) {
-        t_writestring("mkfs: drive not accessible\n");
-        return;
-    }
-
-    int part_idx = part_num - 1;   /* convert to 0-based array index */
-    if (part_idx < 0 || part_idx >= s_cmd_parts.count) {
-        t_writestring("mkfs: invalid partition number");
-        if (s_cmd_parts.count > 0) {
-            t_writestring(" (valid: 1–");
-            t_dec((uint32_t)s_cmd_parts.count);
-            t_putchar(')');
-        }
-        t_writestring("\n      (use lspart ");
-        t_dec(drive);
-        t_writestring(" to list partitions)\n");
-        return;
-    }
-
-    uint32_t lba     = s_cmd_parts.parts[part_idx].lba_start;
-    uint32_t sectors = s_cmd_parts.parts[part_idx].lba_count;
-
-    t_writestring("Formatting drive ");
-    t_dec(drive);
-    t_writestring(" partition ");
-    t_dec((uint32_t)part_num);
-    t_writestring(" (");
-    t_dec(sectors / 2048u);
-    t_writestring(" MiB) as FAT32...\n");
-
-    err = fat32_mkfs(drive, lba, sectors);
-    if (err == -6) {
-        t_writestring("mkfs: partition too small for FAT32 (need >= 32 MiB)\n");
-        return;
-    }
-    if (err) {
-        t_writestring("mkfs: I/O error (");
-        t_dec((uint32_t)(-err));
-        t_writestring(")\n");
-        return;
-    }
-
-    t_writestring("Done.  Mount with: mount ");
-    t_dec(drive);
-    t_putchar(' ');
-    t_dec((uint32_t)part_num);
-    t_putchar('\n');
-}
-
-/* ---------------------------------------------------------------------------
- * ISO9660 commands
- * --------------------------------------------------------------------------- */
-
-/* isols <drive> [path] — list a directory on an ISO9660 CD-ROM. */
-void cmd_isols(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: isols <drive> [path]\n");
-        return;
-    }
-
-    uint8_t drive = (uint8_t)parse_uint(argv[1]);
-    const char *path = (argc >= 3) ? argv[2] : "/";
-
-    int err = iso9660_ls(drive, path);
-    if (err == -2)
-        t_writestring("isols: path not found or not ISO9660\n");
-    else if (err)
-        t_writestring("isols: I/O error\n");
-}
-
-/* ---------------------------------------------------------------------------
- * Installer command
- * --------------------------------------------------------------------------- */
-
-void cmd_install(void)
-{
-    installer_run();
-}
-
-/* ---------------------------------------------------------------------------
- * vics — interactive text editor
- * --------------------------------------------------------------------------- */
-
-void cmd_vics(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: vics <filename>\n");
-        t_writestring("  Opens <filename> for editing. Creates the file on save if it does not exist.\n");
-        t_writestring("  Key bindings: Arrow keys navigate | Ctrl+S save | Ctrl+Q quit\n");
-        return;
-    }
-
-    /* Resolve a bare filename against the current working directory. */
-    const char *arg  = argv[1];
-    const char *cwd  = vfs_getcwd();
-
-    static char full_path[VFS_PATH_MAX];
-
-    if (arg[0] == '/') {
-        /* Absolute path — use as-is. */
-        strncpy(full_path, arg, VFS_PATH_MAX - 1);
-        full_path[VFS_PATH_MAX - 1] = '\0';
-    } else {
-        /* Relative path — prepend CWD. */
-        size_t cwd_len = strlen(cwd);
-        size_t arg_len = strlen(arg);
-        if (cwd_len + 1 + arg_len >= VFS_PATH_MAX) {
-            t_writestring("vics: path too long\n");
-            return;
-        }
-        size_t off = 0;
-        memcpy(full_path, cwd, cwd_len); off += cwd_len;
-        if (cwd[cwd_len - 1] != '/')
-            full_path[off++] = '/';
-        memcpy(full_path + off, arg, arg_len + 1);
-    }
-
-    vics_edit(full_path);
-}
-
-/* ---------------------------------------------------------------------------
- * eject — unmount and eject a storage device.
- *
- * Usage:
- *   eject hdd    — flush and unmount the FAT32 volume mounted at /hd
- *   eject cdrom  — unmount /cdrom and send the ATAPI eject command so the
- *                  tray opens (works in QEMU and on real drives with a
- *                  software-controllable tray)
- * --------------------------------------------------------------------------- */
-
-void cmd_eject(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: eject hdd|cdrom\n");
-        return;
-    }
-
-    const char *target = argv[1];
-
-    if (strcmp(target, "hdd") == 0) {
-        if (!fat32_mounted()) {
-            t_writestring("eject: no HDD volume is mounted\n");
-            return;
-        }
-        fat32_unmount();
-        vfs_notify_hd_unmounted();
-        t_writestring("HDD volume unmounted.\n");
-        return;
-    }
-
-    if (strcmp(target, "cdrom") == 0) {
-        /* Find the ATAPI drive that was registered by vfs_init(). */
-        int cd_drive = -1;
-        for (int i = 0; i < IDE_MAX_DRIVES; i++) {
-            const ide_drive_t *d = ide_get_drive((uint8_t)i);
-            if (d && d->present && d->type == IDE_TYPE_ATAPI) {
-                cd_drive = i;
-                break;
-            }
-        }
-
-        if (cd_drive < 0) {
-            t_writestring("eject: no CD-ROM drive detected\n");
-            return;
-        }
-
-        /* Deregister /cdrom in the VFS before sending the hardware command. */
-        vfs_notify_cdrom_ejected();
-
-        int err = ide_eject_atapi((uint8_t)cd_drive);
-        if (err) {
-            t_writestring("eject: ATAPI eject command failed (err ");
-            t_dec((uint32_t)err);
-            t_writestring(")\n");
-        } else {
-            t_writestring("CD-ROM ejected.\n");
-        }
-        return;
-    }
-
-    t_writestring("eject: unknown target '");
-    t_writestring(target);
-    t_writestring("' — use 'hdd' or 'cdrom'\n");
-}
-
-/* ---------------------------------------------------------------------------
- * exec – load and run an ELF32 executable from the VFS
- *
- * Usage: exec <path>
- *   path: absolute or CWD-relative VFS path to an ELF32 ET_EXEC binary.
- *
- * The binary runs in a fresh ring-3 address space.  All PT_LOAD segments must
- * reside above 0x10000000 (the kernel 256 MiB boundary).
- * --------------------------------------------------------------------------- */
-
-static char s_exec_path[VFS_PATH_MAX];
-
-static void exec_task_entry(void)
-{
-    elf_exec(s_exec_path);
-    task_exit();
-}
-
-void cmd_exec(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: exec <path>\n");
-        return;
-    }
-
-    strncpy(s_exec_path, argv[1], VFS_PATH_MAX - 1);
-    s_exec_path[VFS_PATH_MAX - 1] = '\0';
-
-    task_t *t = task_create("exec", exec_task_entry);
-    if (!t) {
-        t_writestring("exec: task_create failed (pool full?)\n");
-        return;
-    }
-
-    /* Wait for the child to finish before returning to the shell prompt.
-     * Without this, both the exec task and the shell end up in keyboard_getchar
-     * simultaneously and the shell consumes the user's input. */
-    while (t->state != TASK_DEAD)
-        task_yield();
-}
-
-/* ---------------------------------------------------------------------------
- * File I/O commands
- * --------------------------------------------------------------------------- */
-
-/*
- * write <file> <text...>
- *
- * Create or overwrite <file> with the given text (argv[2..] joined by spaces,
- * with a trailing newline).  FAT32 /hd/ paths only.
- */
-void cmd_write(int argc, char **argv)
-{
-    if (argc < 3) {
-        t_writestring("Usage: write <file> <text...>\n");
-        return;
-    }
-
-    static char s_write_buf[SHELL_MAX_INPUT];
-    size_t off = 0;
-
-    for (int i = 2; i < argc; i++) {
-        if (i > 2 && off < sizeof(s_write_buf) - 1)
-            s_write_buf[off++] = ' ';
-        const char *s = argv[i];
-        while (*s && off < sizeof(s_write_buf) - 1)
-            s_write_buf[off++] = *s++;
-    }
-    if (off < sizeof(s_write_buf) - 1)
-        s_write_buf[off++] = '\n';
-    s_write_buf[off] = '\0';
-
-    /* Resolve path against CWD if relative. */
-    static char s_write_path[VFS_PATH_MAX];
-    const char *arg = argv[1];
-    const char *cwd = vfs_getcwd();
-    if (arg[0] == '/') {
-        strncpy(s_write_path, arg, VFS_PATH_MAX - 1);
-        s_write_path[VFS_PATH_MAX - 1] = '\0';
-    } else {
-        size_t cl = strlen(cwd), al = strlen(arg);
-        if (cl + 1 + al >= VFS_PATH_MAX) { t_writestring("write: path too long\n"); return; }
-        size_t p = 0;
-        memcpy(s_write_path, cwd, cl); p += cl;
-        if (cwd[cl - 1] != '/') s_write_path[p++] = '/';
-        memcpy(s_write_path + p, arg, al + 1);
-    }
-
-    int err = vfs_write_file(s_write_path, s_write_buf, (uint32_t)off);
-    if (err) {
-        t_writestring("write: error ");
-        t_dec((uint32_t)(-err));
-        t_putchar('\n');
-    }
-}
-
-/*
- * touch <file>
- *
- * Create an empty file, or do nothing if it already exists (zero-byte write).
- */
-void cmd_touch(int argc, char **argv)
-{
-    if (argc < 2) {
-        t_writestring("Usage: touch <file>\n");
-        return;
-    }
-
-    static char s_touch_path[VFS_PATH_MAX];
-    const char *arg = argv[1];
-    const char *cwd = vfs_getcwd();
-    if (arg[0] == '/') {
-        strncpy(s_touch_path, arg, VFS_PATH_MAX - 1);
-        s_touch_path[VFS_PATH_MAX - 1] = '\0';
-    } else {
-        size_t cl = strlen(cwd), al = strlen(arg);
-        if (cl + 1 + al >= VFS_PATH_MAX) { t_writestring("touch: path too long\n"); return; }
-        size_t p = 0;
-        memcpy(s_touch_path, cwd, cl); p += cl;
-        if (cwd[cl - 1] != '/') s_touch_path[p++] = '/';
-        memcpy(s_touch_path + p, arg, al + 1);
-    }
-
-    int err = vfs_write_file(s_touch_path, "", 0);
-    if (err) {
-        t_writestring("touch: error ");
-        t_dec((uint32_t)(-err));
-        t_putchar('\n');
-    }
-}
-
-/* 64 KiB staging buffer shared by cp (static → lives in BSS, not on stack). */
-static uint8_t s_cp_buf[64u * 1024u];
-
-/*
- * cp <src> <dst>
- *
- * Copy a file.  src may be on /cdrom or /hd; dst must be on /hd (FAT32).
- * Paths are resolved against the CWD if relative.
- */
-void cmd_cp(int argc, char **argv)
-{
-    if (argc < 3) {
-        t_writestring("Usage: cp <src> <dst>\n");
-        return;
-    }
-
-    /* Helper: resolve path (absolute or relative to CWD) into out[outsz]. */
-    static char s_cp_src[VFS_PATH_MAX];
-    static char s_cp_dst[VFS_PATH_MAX];
-
-    const char *cwd = vfs_getcwd();
-
-    const char *paths[2] = { argv[1], argv[2] };
-    char       *outs [2] = { s_cp_src, s_cp_dst };
-
-    for (int i = 0; i < 2; i++) {
-        const char *arg = paths[i];
-        char       *out = outs[i];
-        if (arg[0] == '/') {
-            strncpy(out, arg, VFS_PATH_MAX - 1);
-            out[VFS_PATH_MAX - 1] = '\0';
-        } else {
-            size_t cl = strlen(cwd), al = strlen(arg);
-            if (cl + 1 + al >= VFS_PATH_MAX) { t_writestring("cp: path too long\n"); return; }
-            size_t p = 0;
-            memcpy(out, cwd, cl); p += cl;
-            if (cwd[cl - 1] != '/') out[p++] = '/';
-            memcpy(out + p, arg, al + 1);
-        }
-    }
-
-    uint32_t size = 0;
-    int err = vfs_read_file(s_cp_src, s_cp_buf, sizeof(s_cp_buf), &size);
-    if (err) {
-        t_writestring("cp: cannot read '");
-        t_writestring(s_cp_src);
-        t_writestring("'\n");
-        return;
-    }
-
-    err = vfs_write_file(s_cp_dst, s_cp_buf, size);
-    if (err) {
-        t_writestring("cp: cannot write '");
-        t_writestring(s_cp_dst);
-        t_writestring("'\n");
-        return;
-    }
-
-    t_dec(size);
-    t_writestring(" bytes copied.\n");
-}
